feat : 급상승 선물 랭킹의 아이템들이 그리드 목록으로 나오도록 구현
